import export in control
- Import (aggiornati i visitor per l’importazione aggiungendo controlli sui campi, sostituendo raw pointer con shared ptr, aggiunto l’opzione in parse opening per giornate chiuse, aggiunta eccezione roll back per tipo importato sconosciuto, fatto import automatico)

- Collegato export in xml e json al control

+++ model/Cafe.h
@@ -14,7 +14,7 @@ public:
     Cafe(const QString& name,
          const QString& city,
          const QString& description,
-         float rating,
+         double rating,
          const weeklyOpenings& hours,
          double cost,
          bool takeAway,
