+++ model/Culture.cpp
@@ -1,7 +1,7 @@
 #include "Culture.h"
 
 Culture::Culture(const QString& name, const QString& city, const QString& description,
-                 float rating, const weeklyOpenings& open, double cost,
+                 double rating, const weeklyOpenings& open, double cost,
                  double studentDiscount, bool guidedTour,
                  const QString& culturalFocus, const QString& hasGuidedTour)
     : Place(name, city, description, rating, open, cost),
