+++ model/Disco.cpp
@@ -2,7 +2,7 @@
 #include "visitor/placevisitorinterface.h"
 
 Disco::Disco(const QString& name, const QString& city, const QString& description,
-             float rating, const weeklyOpenings& open, double cost,
+             double rating, const weeklyOpenings& open, double cost,
              double avgStayDuration, int minimumAge, const QString& restrictedEntry,
              const QString& musicGenre, bool hasPrive, const QString& dressCode)
     : Entertainment(name, city, description, rating, open, cost,
