+++ model/Disco.h
@@ -11,7 +11,7 @@ private:
 
 public:
     Disco(const QString& name, const QString& city, const QString& description,
-          float rating, const weeklyOpenings& open, double cost,
+          double rating, const weeklyOpenings& open, double cost,
           double avgStayDuration, int minimumAge, const QString& restrictedEntry,
           const QString& musicGenre, bool hasPrive, const QString& dressCode);
