+++ model/Monument.h
@@ -11,7 +11,7 @@ private:
 
 public:
     Monument(const QString& name, const QString& city, const QString& description,
-             float rating, const weeklyOpenings& open, double cost,
+             double rating, const weeklyOpenings& open, double cost,
              double studentDiscount, bool guidedTour,
              const QString& culturalFocus, const QString& hasGuidedTour,
              bool isUnescoHeritage, const QString& conservationStatus, bool openToPublic);
