+++ model/Restaurant.cpp
@@ -5,7 +5,7 @@
 Restaurant::Restaurant( const QString& name,
                         const QString& city,
                         const QString& description,
-                        float rating,
+                        double rating,
                         const weeklyOpenings& hours,
                         double cost,
                         bool takeAway,
